k8s/requirements.txt
streamlit
spacy
google-cloud-storage
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz